fix: season validator -> time validator
- 해당 클래스에 기능을 추가하면서 이름을 변경했습니다
- 2차 양도 기간이 지났는지 체크하는 로직이 추가되었습니다
